Initial rough draft
Non working version to get the main skeleton of the code done. Utilizing object oriented programming to determine logic for the states and events. Goal is to first get the usages for the magnetometer so it can influence behavior of events. Data can be grabbed from the magnetometer, but I am not sure what to do with the numbers.