working initial versions for uart
These files work between the Feather M4 and the RPi. This issue was changing the byte size to read, causing the struct.unpack method to attempt to read too many bytes to unpack. Timing might be slow for frequency of data s well as clock rate to increase bitrate for real time tracking